- Restruturação do comando gerar para algo mais organizado; - Atualização do pacote `@designliquido/flexoes`.
